+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Fixes:
 
+* Fix subdirectory config url
 * Fix google2fa column size
 * Fix errors display for api
 * Fix currency in double